Support application keys and developer credentials APIs
